Second sweep of reorg and refactoring
