+++ DNN Platform/DotNetNuke.Abstractions/Prompt/ICommandHelp.cs
@@ -1,4 +1,8 @@
-﻿using System.Collections.Generic;
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+using System.Collections.Generic;
 
 namespace DotNetNuke.Abstractions.Prompt
 {
