--- DNN Platform/DotNetNuke.Abstractions/Prompt/ICommandInputModel.cs
@@ -1,9 +1,0 @@
-﻿namespace DotNetNuke.Abstractions.Prompt
-{
-    public interface ICommandInputModel
-    {
-        string[] Args { get; }
-        string CmdLine { get; set; }
-        int CurrentPage { get; set; }
-    }
-}
