+++ DNN Platform/DotNetNuke.Abstractions/Prompt/ICommandOption.cs
@@ -1,4 +1,8 @@
-﻿namespace DotNetNuke.Abstractions.Prompt
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+namespace DotNetNuke.Abstractions.Prompt
 {
     public interface ICommandOption
     {
