+++ DNN Platform/DotNetNuke.Abstractions/Prompt/IConsoleCommand.cs
@@ -2,7 +2,6 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // 
-
 using DotNetNuke.Abstractions.Users;
 
 namespace DotNetNuke.Abstractions.Prompt
