+++ DNN Platform/Library/DotNetNuke.Library.csproj
@@ -250,15 +250,15 @@
     <Compile Include="Entities\Users\UserRegistrationEmailNotifier.cs" />
     <Compile Include="Obsolete\PortalSettings.cs" />
     <Compile Include="Prompt\Command.cs" />
-    <Compile Include="Prompt\CommandHelp.cs" />
-    <Compile Include="Prompt\CommandOption.cs" />
+    <Compile Include="Prompt\Output\CommandHelp.cs" />
+    <Compile Include="Prompt\Output\CommandOption.cs" />
     <Compile Include="Prompt\CommandRepository.cs" />
     <Compile Include="Prompt\Common\Constants.cs" />
     <Compile Include="Prompt\Attributes\ConsoleCommandParameterAttribute.cs" />
     <Compile Include="Prompt\Attributes\ConsoleCommandAttribute.cs" />
     <Compile Include="Prompt\ConsoleCommand.cs" />
-    <Compile Include="Prompt\ConsoleErrorResultModel.cs" />
-    <Compile Include="Prompt\ConsoleResultModel.cs" />
+    <Compile Include="Prompt\Output\ConsoleErrorResultModel.cs" />
+    <Compile Include="Prompt\Output\ConsoleResultModel.cs" />
     <Compile Include="Prompt\Common\PagingInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Collections\CollectionExtensions.cs" />
