+++ DNN Platform/Library/Prompt/Attributes/ConsoleCommandAttribute.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+using System;
 
 namespace DotNetNuke.Prompt.Attributes
 {
