+++ DNN Platform/Library/Prompt/Command.cs
@@ -1,4 +1,8 @@
-﻿using DotNetNuke.Abstractions.Prompt;
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+using DotNetNuke.Abstractions.Prompt;
 using Newtonsoft.Json;
 using System;
