+++ DNN Platform/Library/Prompt/CommandRepository.cs
@@ -1,8 +1,13 @@
-﻿using DotNetNuke.Abstractions.Prompt;
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+using DotNetNuke.Abstractions.Prompt;
 using DotNetNuke.Common.Utilities;
 using DotNetNuke.Framework;
 using DotNetNuke.Framework.Reflections;
 using DotNetNuke.Prompt.Attributes;
+using DotNetNuke.Prompt.Output;
 using DotNetNuke.Services.Localization;
 using System;
 using System.Collections.Generic;
