+++ DNN Platform/Library/Prompt/Common/Constants.cs
@@ -1,4 +1,8 @@
-﻿namespace DotNetNuke.Prompt.Common
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+namespace DotNetNuke.Prompt.Common
 {
     public static class Constants
     {
