+++ DNN Platform/Library/Prompt/Output/CommandHelp.cs
@@ -1,9 +1,13 @@
-﻿using DotNetNuke.Abstractions.Prompt;
+﻿// 
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// 
+using DotNetNuke.Abstractions.Prompt;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 
-namespace DotNetNuke.Prompt
+namespace DotNetNuke.Prompt.Output
 {
     [Serializable]
     [JsonObject]
