+++ DNN Platform/Library/Prompt/Output/ConsoleErrorResultModel.cs
@@ -2,19 +2,16 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // 
-using System.IO;
-using Localization = DotNetNuke.Services.Localization.Localization;
+using DotNetNuke.Services.Localization;
 
-namespace DotNetNuke.Prompt
+namespace DotNetNuke.Prompt.Output
 {
     public class ConsoleErrorResultModel : ConsoleResultModel
     {
-        private static string LocalResourcesFile => Path.Combine("~/DesktopModules/admin/Dnn.PersonaBar/App_LocalResources/SharedResources.resx");
-
         public ConsoleErrorResultModel()
         {
             IsError = true;
-            Output = Localization.GetString("Prompt_InvalidSyntax", LocalResourcesFile, true);
+            Output = Localization.GetString("Prompt_InvalidSyntax", Common.Constants.DefaultPromptResourceFile, true);
         }
 
         public ConsoleErrorResultModel(string errMessage)
